backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
aiosqlite==0.21.0
aiofiles==24.1.0
pydantic==2.11.9
pyserial==3.5
